app/pyproject.toml
[tool.poetry]
name = "ikapi"
version = "0.1.0"
description = ""
authors = ["maasy"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.59.0"
uvicorn = "^0.11.5"
tortoise-orm = "^0.16.13"
aiomysql = "^0.0.20"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
